requirements.txt
Django==5.2.*
environs==14.2.*
dj-database-url==3.0.*
django-phonenumber-field[phonenumbers]==7.0.1
phonenumbers==8.13.17